requirements.txt
pandas>=2.1.0
numpy>=1.25.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
seaborn>=0.13.0
Flask>=3.0.0
joblib>=1.3.0
python-dotenv>=1.0.0
pytest>=7.4.0
PyYAML>=6.0.0
